# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AnimeStatisticsResponse": {
        "properties": {
          "count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count"
          },
          "episodesWatched": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episodeswatched"
          },
          "meanScore": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meanscore"
          },
          "standardDeviation": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Standarddeviation"
          }
        },
        "title": "AnimeStatisticsResponse",
        "type": "object"
      },
...
      "AvatarResponse": {
        "description": "Avatar urls shared by profile, activities and replies payloads.",
        "properties": {
          "large": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Large"
          },
          "medium": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medium"
          }
        },
        "title": "AvatarResponse",
        "type": "object"
      },
...
      "DeleteReplyResponse": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          }
        },
        "required": [
          "deleted"
        ],
        "title": "DeleteReplyResponse",
        "type": "object"
      },
      "ErrorCode": {
        "enum": [
          "BAD_REQUEST",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "NOT_FOUND",
          "VALIDATION_ERROR",
          "INTERNAL_ERROR",
          "UPSTREAM_ERROR"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetail": {
        "description": "Field-level issue attached to :class:`ErrorResponse` (422 responses).",
        "properties": {
          "loc": {
            "title": "Loc",
            "type": "string"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "Body of every error response produced by the API.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "title": "Errors",
            "type": "array"
          }
        },
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "LikeableType": {
        "description": "AniList ``LikeableType`` — target accepted by the ToggleLike mutation.",
        "enum": [
          "THREAD",
          "THREAD_COMMENT",
          "ACTIVITY",
          "ACTIVITY_REPLY"
        ],
        "title": "LikeableType",
        "type": "string"
      },
...
      "MeResponse": {
        "description": "Authenticated user, in a single shape for both cases below.\n\n``exists`` is ``False`` when the app has a valid JWT but no matching row in\n``users`` (the token was issued before the account row was removed).",
        "properties": {
          "anilist_id": {
            "title": "Anilist Id",
            "type": "integer"
          },
          "exists": {
            "title": "Exists",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "required": [
          "id",
          "anilist_id",
          "exists"
        ],
        "title": "MeResponse",
        "type": "object"
      },
...
      "StatisticsResponse": {
        "properties": {
          "anime": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnimeStatisticsResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "StatisticsResponse",
        "type": "object"
      },
...
      "ToggleLikeResponse": {
        "description": "Result of liking/unliking an activity or a reply.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "isLiked": {
            "default": false,
            "title": "Isliked",
            "type": "boolean"
          },
          "likeCount": {
            "default": 0,
            "title": "Likecount",
            "type": "integer"
          }
        },
        "required": [
          "id"
        ],
        "title": "ToggleLikeResponse",
        "type": "object"
      },
...
      "UserProfileResponse": {
        "description": "Public profile of the authenticated AniList account (``/me/profile``).",
        "properties": {
          "about": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "About"
          },
          "avatar": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AvatarResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "bannerImage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bannerimage"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "statistics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StatisticsResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "UserProfileResponse",
        "type": "object"
      },
      "ViewerMediaListOptions": {
        "properties": {
          "scoreFormat": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scoreformat"
          }
        },
        "title": "ViewerMediaListOptions",
        "type": "object"
      },
      "ViewerResponse": {
        "description": "The AniList account behind the current app session (``/me/viewer``).",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "mediaListOptions": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ViewerMediaListOptions"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "ViewerResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Neko Track Backend",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/me/profile": {
      "get": {
        "operationId": "profile_api_v1_me_profile_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid app JWT"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "AniList account not connected"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid AniList response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Profile",
        "tags": [
          "anilist",
          "profile"
        ]
      }
    },
    "/api/v1/me/viewer": {
      "get": {
        "operationId": "get_viewer_api_v1_me_viewer_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid app JWT"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "AniList account not connected"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid AniList response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Viewer",
        "tags": [
          "anilist",
          "profile"
        ]
      }
    },
    "/api/v1/replies/{reply_id}": {
      "delete": {
        "operationId": "delete_reply_api_v1_replies__reply_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "reply_id",
            "required": true,
            "schema": {
              "title": "Reply Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteReplyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid app JWT"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "AniList account not connected"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid AniList response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Reply",
        "tags": [
          "replies",
          "activities"
        ]
      }
    },
    "/api/v1/replies/{reply_id}/toggle-like": {
      "post": {
        "operationId": "toggle_reply_like_api_v1_replies__reply_id__toggle_like_post",
        "parameters": [
          {
            "in": "path",
            "name": "reply_id",
            "required": true,
            "schema": {
              "title": "Reply Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/LikeableType"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToggleLikeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid app JWT"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "AniList account not connected"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid AniList response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Toggle Reply Like",
        "tags": [
          "replies",
          "activities"
        ]
      }
    },
    "/auth/anilist/callback": {
      "get": {
        "operationId": "callback_auth_anilist_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "title": "State",
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to the app deep link carrying the JWT"
          },
          "400": {
            "description": "Missing, invalid or replayed state"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid app JWT"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "AniList account not connected"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid AniList response"
          }
        },
        "summary": "Callback",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/anilist/me": {
      "get": {
        "operationId": "me_auth_anilist_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid app JWT"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "AniList account not connected"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid AniList response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Me",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/anilist/start": {
      "get": {
        "operationId": "start_auth_anilist_start_get",
        "responses": {
          "302": {
            "description": "Redirect to the AniList authorize URL"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid app JWT"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "AniList account not connected"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid AniList response"
          }
        },
        "summary": "Start",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}